requirements.txt
Django==1.10.6
django-crispy-forms==1.6.1
django-filter==1.0.2
djangorestframework==3.6.2
Markdown==2.6.8
pkg-resources==0.0.0
pytz==2017.2
